[src] remove old version
