+++ Sources/CohesionKit/Visitor/EntityContext.swift
@@ -1,5 +1,6 @@
 import Foundation
 
+/// Information related to an entity while storing it into the `IdentityMap`
 struct EntityContext<Root, Value> {
     let parent: EntityNode<Root>
     let keyPath: KeyPath<Root, Value>
